backend/requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
openpyxl==3.1.2
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1

